Clean up, move, or gut some outdated documents.
The near-term goal is to get rid of the articles folder and work all
ideas into the spec. The next goal after that is to take everything out
of the tex version of the spec and delete that too. Once we have
accomplished this, we finally have ONE, relatively clean version of the
Lore specification.