Remove this association business from Feedback.
Associating fragments with feedback instances after their creation was a
nice idea, since we wouldn't have had to pass the fragments down the
call chain. However, it was simply too opaque which error had already
been associated with a fragment, and so I ultimately decided to make a
Position mandatory for every feedback instance on creation.