Reimplement LUB to properly handle sum types embedded in intersection types.
The trick is forcing the LUB that handles intersection types to never
return a sum type. Then our algorithm for deciding the LUB of
intersection types works. On the other side, we can still fall back to a
sum type LUB if Any is returned, which is now configurable when calling
configurableLub.
Also add a SumType.join function, which attempts to reduce sum types to
their closest collective supertype.