Parts of the spec have been moved to a (currently local) dokuwiki.
I am currently in the process of migrating the specification to
dokuwiki, as it affords much better organization and cross-linking. It
also helps hiding the less user-relevant parts of the spec, such as
formal definitions and proofs, from the average user. This makes the
specification cleaner and more readable. It also allows me to better
separate practical topics (such as syntax and semantics of the language)
from formal topics (e.g. the formal definition of multiple dispatch or
the exact subtyping/specificity rules).