Refactoring: Introduce Node.State, which encapsulates state in one object.
The advantage of this approach is simple: we can copy nodes more easily,
because the state will also be copied. Also, accessign mutable node
state is now explicitly done through node.state, which is very clear.
Right now, copying a node will not deep-copy the State. This is an
oversight which will be corrected soon.
Pull fragments into nodes so that we don't have to pass the fragment
around like crazy. This greatly simplifies the way we handle positions:
Fragments are passed around during the parsing phase (to the respective
parser classes), and past that we only pass full positions. Position is,
once again, a single class. The resolution phase no takes a single huge
list of declarations instead of a list of fragments. Fragments don't
save their own declarations anymore (this was not exactly thread-safe).
Move string parsing to StatementParser, which removes LexicalParser's
dependency on StatementParser.
Remove FragmentNode, which has become obsolete as hell. Thankfully.
Add comparisons.lore.