Use the Expression representation for transformation/transpilation.
This representation is used during transpilation and produced by the
verification/transformation phase. The biggest advantage to this is, of
course, the reclaimed immutability, which makes writing the verification
and transpilation phases much easier. (Node definitions can also be
vastly simplified.) There is also the added advantage that data produced
by the parser and consumed by the transformation phase (and by the time
we transpile obsolete) is not carried to the transpilation phase. This
makes the individual components of the compiler clearner.
Existing StmtTransformers are rolled into the transformation phase.
We remove the need for deferred typings by traversing type declaration
nodes twice: Once for registering the types and then for actually
resolving label/class definitions.
ClassDefinition now receives its constructors later. This allows us to
generate the default constructor during resolution, which in general
improves the clarity surrounding default constructor compilation.
Tests still have to be fixed!