Add type benchmarks for areEqual.
The problem with this code is that V8 doesn't seem to read the Kinds as
constants, and neither does it seem to optimize the switch with a jump
table. This is very disappointing.

However, the performance is still OK. The next commit will contain an
implementation of areEqual in Rust. This implementation is about 2-3
times faster than the JS code. I would say it's unfortunate that the JS
code is so slow, but it isn't MAGNITUDES slower.

Obviously, JS is a great testbed for prototyping the language, so I'm
not even thinking about switching to any sort of Rust-based runtime
system anytime soon. There is still a lot of potential for optimization
with partial dispatch (some functions aren't subtyped in all of their
parameters), calling dispatch functions that start deeper into the tree
instead of at the root, calling functions directly because they only
have one dispatch option, and so on. We can implement these options in
the JS code generated by the compiler. They will probably have a bigger
impact than switching to a faster runtime system.

Some day, however, there will come a point when it'd be beneficial to
switch to a faster (or at least cleaner) run-time. Then it might be time
to implement a VM in Rust (that can natively support multiple-dispatch)
or even to compile Lore with LLVM AOT.