Implement areEqual using a "lambda jump table" for kinds.
Implementing the algorithm with a switch on direct integers (as opposed
to Kind.XY variables or constants) was technically a little bit faster,
but very ugly and not worth the performance improvement for now. The
lambda jump method is surprisingly fast (faster than switching on the
kind with the original implementation) and even an improvement from a
code quality standpoint.