Refactoring: Vector replaces List as the default sequence.
Lists are still used to represent queues. The main rationale for using
vectors is that they are the better general purpose data structure,
making it harder to commit performance crimes. Vectors also support
appending in a performant manner, which is more natural to me than
prepending and sometimes reversing. Lists may be faster in some
scenarios, but the compiler is not performance-critical and will be
optimized as needed.

These changes have not yet been applied to the verification and
transpilation phases.