Update monster example with struct/trait changes.
The way the name property is carried across the type hierarchy isn't
quite nice mostly in a syntactic way and from a standpoint of "don't
repeat yourself". The former can be solved by adding syntactic sugar for
trait "properties", while the latter could be solved by a mixin system,
which is still under long-term consideration.