Add basic combat simulation test runner.
The test simply checks that the winner is the radiant. This might be
expanded in the future.
Fix: When the test code prints messages, the JSON result wasn't
addressed correctly. Because the JSON result will always be the last
logged message, we can simply address that particular one instead of the
first message.